remove dead handlers and move specific handlers to point of use
