+++ pkg/cmd/server/origin/handlers.go
@@ -1,17 +1,30 @@
 package origin
 
 import (
+	"bytes"
+	"encoding/json"
+	"errors"
+	"fmt"
 	"net/http"
 	"regexp"
 	"strings"
 
+	restful "github.com/emicklei/go-restful"
 	"github.com/golang/glog"
 
+	authenticationv1 "k8s.io/api/authentication/v1"
+	kapierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/util/sets"
+	kauthorizer "k8s.io/apiserver/pkg/authorization/authorizer"
 	apirequest "k8s.io/apiserver/pkg/endpoints/request"
+	"k8s.io/kubernetes/pkg/api/legacyscheme"
+	coreapi "k8s.io/kubernetes/pkg/apis/core"
 
+	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
 	configapi "github.com/openshift/origin/pkg/cmd/server/api"
-	serverhandlers "github.com/openshift/origin/pkg/cmd/server/handlers"
+	authenticationapi "github.com/openshift/origin/pkg/oauthserver/api"
 )
 
 // cacheExcludedPaths is small and simple until the handlers include the cache headers they need
@@ -38,38 +51,6 @@ func cacheControlFilter(handler http.Handler, value string) http.Handler {
 		}
 
 		w.Header().Set("Cache-Control", value)
-		handler.ServeHTTP(w, req)
-	})
-}
-
-// namespacingFilter adds a filter that adds the namespace of the request to the context.  Not all requests will have namespaces,
-// but any that do will have the appropriate value added.
-func namespacingFilter(handler http.Handler, contextMapper apirequest.RequestContextMapper) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		ctx, ok := contextMapper.Get(req)
-		if !ok {
-			http.Error(w, "Unable to find request context", http.StatusInternalServerError)
-			return
-		}
-
-		if _, exists := apirequest.NamespaceFrom(ctx); !exists {
-			if requestInfo, ok := apirequest.RequestInfoFrom(ctx); ok && requestInfo != nil {
-				// only set the namespace if the apiRequestInfo was resolved
-				// keep in mind that GetAPIRequestInfo will fail on non-api requests, so don't fail the entire http request on that
-				// kind of failure.
-
-				// TODO reconsider special casing this.  Having the special case hereallow us to fully share the kube
-				// APIRequestInfoResolver without any modification or customization.
-				namespace := requestInfo.Namespace
-				if (requestInfo.Resource == "projects") && (len(requestInfo.Name) > 0) {
-					namespace = requestInfo.Name
-				}
-
-				ctx = apirequest.WithNamespace(ctx, namespace)
-				contextMapper.Update(req, ctx)
-			}
-		}
-
 		handler.ServeHTTP(w, req)
 	})
 }
@@ -158,18 +139,65 @@ func (c *MasterConfig) versionSkewFilter(handler http.Handler, contextMapper api
 			}
 
 			if !foundMatch {
-				serverhandlers.Forbidden(defaultMessage, nil, w, req)
+				forbidden(defaultMessage, nil, w, req)
 				return
 			}
 		}
 
 		for _, filter := range deniedFilters {
 			if filter.matches(req.Method, userAgent) {
-				serverhandlers.Forbidden(filter.message, nil, w, req)
+				forbidden(filter.message, nil, w, req)
 				return
 			}
 		}
 
 		handler.ServeHTTP(w, req)
 	})
+}
+
+// translateLegacyScopeImpersonation is a filter that will translates user scope impersonation for openshift into the equivalent kube headers.
+func translateLegacyScopeImpersonation(handler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		for _, scope := range req.Header[authenticationapi.ImpersonateUserScopeHeader] {
+			req.Header[authenticationv1.ImpersonateUserExtraHeaderPrefix+authorizationapi.ScopesKey] =
+				append(req.Header[authenticationv1.ImpersonateUserExtraHeaderPrefix+authorizationapi.ScopesKey], scope)
+		}
+
+		handler.ServeHTTP(w, req)
+	})
+}
+
+// forbidden renders a simple forbidden error to the response
+func forbidden(reason string, attributes kauthorizer.Attributes, w http.ResponseWriter, req *http.Request) {
+	resource := ""
+	group := ""
+	name := ""
+	// the attributes can be empty for two basic reasons:
+	// 1. malformed API request
+	// 2. not an API request at all
+	// In these cases, just assume default that will work better than nothing
+	if attributes != nil {
+		group = attributes.GetAPIGroup()
+		resource = attributes.GetResource()
+		name = attributes.GetName()
+	}
+
+	// Reason is an opaque string that describes why access is allowed or forbidden (forbidden by the time we reach here).
+	// We don't have direct access to kind or name (not that those apply either in the general case)
+	// We create a NewForbidden to stay close the API, but then we override the message to get a serialization
+	// that makes sense when a human reads it.
+	forbiddenError := kapierrors.NewForbidden(schema.GroupResource{Group: group, Resource: resource}, name, errors.New("") /*discarded*/)
+	forbiddenError.ErrStatus.Message = reason
+
+	formatted := &bytes.Buffer{}
+	output, err := runtime.Encode(legacyscheme.Codecs.LegacyCodec(coreapi.SchemeGroupVersion), &forbiddenError.ErrStatus)
+	if err != nil {
+		fmt.Fprintf(formatted, "%s", forbiddenError.Error())
+	} else {
+		json.Indent(formatted, output, "", "  ")
+	}
+
+	w.Header().Set("Content-Type", restful.MIME_JSON)
+	w.WriteHeader(http.StatusForbidden)
+	w.Write(formatted.Bytes())
 }
